+++ libs/go-restful/container.go
@@ -14,7 +14,7 @@ import (
 	"strings"
 	"sync"
 
-	"github.com/emicklei/go-restful/log"
+	"github.com/spring1843/chat-server/libs/go-restful/log"
 )
 
 // Container holds a collection of WebServices and a http.ServeMux to dispatch http requests.