+++ libs/go-restful/logger.go
@@ -4,7 +4,7 @@ package restful
 // Use of this source code is governed by a license
 // that can be found in the LICENSE file.
 import (
-	"github.com/emicklei/go-restful/log"
+	"github.com/spring1843/chat-server/libs/go-restful/log"
 )
 
 var trace bool = false